rename file, add paser data, change name method and variables with case deletion, delete old methods, create method to convert altaz in ardec
